Final version of platform Game on private repo
Final version of platform Game on private repo

+++ Platform Game/Assets/Scripts/EnemyScriptableObject.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class EnemyScriptableObject : ScriptableObject
 {
     public float targetDistance = 15f;
-    public float specialAttackDuration = 3f;
+    public int specialAttackDuration = 3;
 }
